requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
DateTime==4.2
Django==1.11.5
djangorestframework==3.11.1
docopt==0.6.2
idna==2.6
Jinja2==2.9.6
MarkupSafe==1.1.1
netCDF4==1.2.9
numpy==1.13.1
pandas==0.20.3
pip==9.0.1
Pydap==3.2.2
pytz==2017.2
requests==2.18.4
setuptools==28.8.0
six==1.11.0
urllib3==1.22
WebOb==1.7.3
wfastcgi==3.0.0
xlwt==1.3.0
zope.interface==4.4.3
pysofar>=0.1.8
